Added dynamic ControlPath definition
In order to solve the problem of unnecessary `LocalForward` definitions
when creating multiple connections to the same host, a scenario that
occurs when a control path is used, specifying the `ControlPath` dynamically
allows detection of an active connection. When the first connection is created
the `ControlPath` is generated by SSH and we save the ports in the cache
directory. For the next connection, if the `Controlpath` is still in existence
then we can specify identical `LocalForward` entries without an issue.

* Updated `cfg.EnvSSHIdentityFile` to be the full, default path for an
  ED25519 private key
* Added `ssh.setControlPath` to dynamically define the `ControlPath` for
  use with the connection
* Updated `ssh.setPortForwarding` to check for a `ControlPath` file and,
  when it exists along with a cached set of ports, reuse the ports. If
  we don't have what we expect, or there is no cache for the ports, we
  reset and generate new ones and set the cache
* Added `ssh.TestConnection`
* Fixed simplication issue with reflect