Hash ControlPath socket names by default (#66)
Currently, the dynamic ControlPath is done in such a way as to make it easy to determine its purpose. However, should long HostName fields exist then this could potentially exceed the maximum path length for a UNIX socket (UNIX_PATH_MAX).

By switching to using a hash, similar to %C in ssh, we can restrict the length of the path. In order to allow graceful migration across from the old format we can allow either format when determining if a socket is present, but only use hashed path names for new ones.

Addresses #65

* Use hashed ControlPath files by default
* Updated fallback option from '%r@%h:%p' to '%C'
* Added conditional check for existing unhashed paths,
  which allows existing ControlPath sockets to be used
  until expiry. When there isn't an existing socket, a
  sha256 hash is created instead
* Allow toggle for XZ
* Updated tests for ssh package
* Added TestControlPath and moved tests out of TestConnection
* Added dummyArgs and used for tests using ssh.BuildConnection
  instead of an empty map[string]interface{}
* Ensure unique entries in Connection.LocalForward