Add option to view usable placeholders for User
The User option for the SSH connection can be stored as a template
so that it changes for each user, based upon some pre-defined criteria.
Without looking at the code, it is not possible to see the supported
placeholders.

Adding an option for the user to list the available ones makes the use
of templated users easier.

* Added `cmd.inspectItem` to reveal the value of a predefined set of
  terms, currently just placeholders for `User`
* Added `cmd.inspectCmd` to call `inspectItem`
* Updated `README`

Addresses https://github.com/cezmunsta/ssh_ms/issues/67