Refactored unit tests to use containers
The direct dependency on the main Vault package causes a number of
issues, such as unnecessary Dependabot activity due to CVEs in code
the is not used, etc. The SDK contains code that allows a test
cluster to be created from code and thus allows a more specific set of
dependencies and less maintenance.

* Removed dependency of unit tests on github.com/hashicorp/vault and
  github.com/hashicorp/vault-plugin-secrets-kv, reducing the indirect
  dependencies (and thus potential conflicts) in addition.
* Bumped Go version to v1.21
* Added helpers package to allow reuse of the test cluster
* Updated cmd.TestMain to use os.TempDir instead of static path